21.H Favorite button working dinamically
location: http://127.0.0.1:8000/questions/my-question
next to the question there should be a gray star,
if you are log in and clic the star it will turn into yellow and the coutn increse, if you click again it will turn into gray and the count decrest.
if you are not log in, and click the star button if will redirect you to login page.